Expose product views for reuse
